Rename the Scala package
