+++ sierra_adapter/sierra_bibs_to_sns/src/main/scala/uk/ac/wellcome/platform/sierra_bibs_to_dynamo/Server.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_bibs_to_dynamo
+package uk.ac.wellcome.platform.sierra_bibs_to_sns
 
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.HttpServer
@@ -11,13 +11,13 @@ import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.sierra_adapter.modules.SierraDynamoConfigModule
-import uk.ac.wellcome.platform.sierra_bibs_to_dynamo.modules.SierraBibsToDynamoModule
+import uk.ac.wellcome.platform.sierra_bibs_to_sns.modules.SierraBibsToDynamoModule
 
 object ServerMain extends Server
 
 class Server extends HttpServer {
   override val name =
-    "uk.ac.wellcome.platform.sierra_bibs_to_dynamo SierraBibsToDynamo"
+    "uk.ac.wellcome.platform.sierra_bibs_to_sns SierraBibsToDynamo"
   override val modules = Seq(
     SierraBibsToDynamoModule,
     SierraDynamoConfigModule,
