+++ sierra_adapter/sierra_bibs_to_sns/src/test/scala/uk/ac/wellcome/platform/sierra_bibs_to_dynamo/ServerTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_bibs_to_dynamo
+package uk.ac.wellcome.platform.sierra_bibs_to_sns
 
 import com.twitter.finagle.http.Status._
 import com.twitter.finatra.http.EmbeddedHttpServer
