+++ sierra_adapter/sierra_bibs_to_sns/src/test/scala/uk/ac/wellcome/platform/sierra_bibs_to_dynamo/sink/SierraBibsDynamoSinkTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_bibs_to_dynamo.sink
+package uk.ac.wellcome.platform.sierra_bibs_to_sns.sink
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
@@ -13,7 +13,7 @@ import io.circe.parser._
 import org.mockito.Matchers.any
 import org.mockito.Mockito.when
 import org.scalatest.mockito.MockitoSugar
-import uk.ac.wellcome.platform.sierra_bibs_to_dynamo.locals.SierraBibsToDynamoDBLocal
+import uk.ac.wellcome.platform.sierra_bibs_to_sns.locals.SierraBibsToDynamoDBLocal
 import uk.ac.wellcome.dynamo._
 import uk.ac.wellcome.models.transformable.sierra.SierraBibRecord
 import uk.ac.wellcome.test.utils.ExtendedPatience
